Adding logic to switch Area from View mode to Edit mode. Adding Mapbox night layer + cleaning up tile URLs

+++ src/map/DrawingTool.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { Component } from 'react';
 import { EditControl } from "react-leaflet-draw";
 import { FeatureGroup } from 'react-leaflet';
 
@@ -6,32 +6,60 @@ import { drawingBuffer } from '../model/DrawingModel';
 import { fsm } from '../model/UIState';
 
 
-const DrawingTool = () => {
-    return (
-        <FeatureGroup>
-            <EditControl
-                position='topright'
-                onEdited={ _onEdited }
-                onCreated={_onCreate}
-                onDeleted={_onDeleted}
-                onDeleteStop={_onDeleteStop}
-                draw={{
-                    rectangle: false,
-                    circle: false,
-                    polygon: {
-                    shapeOptions: {
-                        color: '#8B4513',
-                        weight: 2,
-                        fillColor: '#778899',
-                        fillOpacity: 0.08,
-                        clickable: false,
-                        lineJoin: 'round'
+class DrawingTool extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            object: null
+        };
+
+        this.featureGroupRef = null;
+        this.editControlRef = null;
+
+    }
+
+    componentDidMount() {
+        console.log('DrawingTool', this.featureGroupRef, this.editControlRef);
+        drawingBuffer.clipboard.observe( (change) => {
+            console.log('DrawingTool detecting new item in clipboard', change.newValue);
+            change.newValue.addTo(this.featureGroupRef.leafletElement);
+            this.editControlRef.props.onCreated.call(this.editControlRef, {layerType: 'polygon', layer: change.newValue});
+            //this.setState(()=> ({object: change.newValue}));
+
+        });
+    }
+
+    addOject(obj) {
+        obj.addTo(this.featureGroupRef.leafletElement);
+    }
+
+    render() {
+        return (
+            <FeatureGroup ref={ (ref) => {this.featureGroupRef = ref}}>
+                <EditControl
+                    ref={ ref => this.editControlRef = ref }
+                    position='topright'
+                    onEdited={ _onEdited }
+                    onCreated={_onCreate}
+                    onDeleted={_onDeleted}
+                    onDeleteStop={_onDeleteStop}
+                    draw={{
+                        rectangle: false,
+                        circle: false,
+                        polygon: {
+                        shapeOptions: {
+                            color: '#8B4513',
+                            weight: 2,
+                            fillColor: '#778899',
+                            fillOpacity: 0.08,
+                            clickable: false,
+                            lineJoin: 'round'
+                            }
                         }
-                    }
-                }}
-            />
-        </FeatureGroup> 
-    );
+                    }}
+                />
+            </FeatureGroup> )
+    }
 }
 
 export default DrawingTool;
@@ -71,15 +99,16 @@ const _onDeleted = (e) => {
 
 
 const _onEdited = e => {
-    console.log('_onEdited ', e);
-    const layer = e.target;
-    e.layers.eachLayer(
+    console.log('_onEdited ', e, e.layerType);
+    e.layers && e.layers.eachLayer(
         layer => {
             console.log('Edited layer: ', layer);
             const id = layer._leaflet_id;
             const obj = drawingBuffer.data.get(id);
-            drawingBuffer.deleteObject(id);
-            drawingBuffer.addObject(obj.layer, obj.type);
+            if (obj !== undefined) {
+                drawingBuffer.deleteObject(id);
+                drawingBuffer.addObject(obj.layer, obj.type);
+            }
         }
     );
